+++ forui/example/lib/sandbox.dart
@@ -11,30 +11,24 @@ class Sandbox extends StatefulWidget {
 
 enum Notification { all, direct, nothing }
 
-class _SandboxState extends State<Sandbox> {
+class _SandboxState extends State<Sandbox> with SingleTickerProviderStateMixin {
   bool value = false;
-  FMultiSelectGroupController<int> selectGroupController = FMultiSelectGroupController(min: 2);
-  FAccordionController controller = FAccordionController(min: 1, max: 3);
+  FRadioSelectGroupController<String> selectGroupController = FRadioSelectGroupController();
+  late FPopoverController controller;
 
   @override
   void initState() {
     super.initState();
+    controller = FPopoverController(vsync: this);
   }
 
   @override
-  Widget build(BuildContext context) => FSelectTileGroup<int>(
-        controller: selectGroupController,
-        autovalidateMode: AutovalidateMode.always,
-        validator: (values) => values?.isEmpty ?? true ? 'error message' : null,
-        children: [
-          FSelectTile.suffix(
-            title: const Text('1'),
-            value: 1,
-          ),
-          FSelectTile.suffix(
-            title: const Text('2'),
-            value: 2,
-          ),
-        ],
-      );
+  Widget build(BuildContext context) => const Column();
+
+  @override
+  void dispose() {
+    controller.dispose();
+    selectGroupController.dispose();
+    super.dispose();
+  }
 }