+++ forui/lib/widgets/popover_menu.dart
@@ -1,0 +1,8 @@
+/// {@category Widgets}
+///
+/// A popover menu displays a menu in a portal aligned to a child.
+///
+/// See https://forui.dev/docs/overlay/popover-menu for working examples.
+library forui.widgets.popover_menu;
+
+export '../src/widgets/popover_menu.dart';