+++ forui/test/src/foundation/portal/portal_golden_test.dart
@@ -67,5 +67,38 @@ void main() {
 
       await expectLater(find.byType(TestScaffold), matchesGoldenFile('portal/shown.png'));
     });
+
+    testWidgets('shifted', (tester) async {
+      final controller = OverlayPortalController();
+
+      await tester.pumpWidget(
+        TestScaffold.app(
+          child: FPortal(
+            controller: controller,
+            offset: const Offset(100, 100),
+            followerBuilder: (context) => const Padding(
+              padding: EdgeInsets.all(5),
+              child: ColoredBox(
+                color: Colors.red,
+                child: SizedBox.square(
+                  dimension: 100,
+                ),
+              ),
+            ),
+            child: const ColoredBox(
+              color: Colors.yellow,
+              child: SizedBox.square(
+                dimension: 50,
+              ),
+            ),
+          ),
+        ),
+      );
+
+      controller.show();
+      await tester.pumpAndSettle();
+
+      await expectLater(find.byType(TestScaffold), matchesGoldenFile('portal/shifted.png'));
+    });
   });
 }