+++ forui/test/src/widgets/popover/popover_test.dart
@@ -6,21 +6,20 @@ import 'package:forui/forui.dart';
 import '../../test_scaffold.dart';
 
 void main() {
+  late FPopoverController controller;
+
+  setUp(() => controller = FPopoverController(vsync: const TestVSync()));
+
   group('FPopover', () {
     testWidgets('tap outside hides popover', (tester) async {
-      final controller = FPopoverController(vsync: const TestVSync());
-
       await tester.pumpWidget(
-        MaterialApp(
-          home: TestScaffold(
-            theme: FThemes.zinc.light,
-            child: FPopover(
-              controller: controller,
-              followerBuilder: (context, style, _) => const Text('follower'),
-              target: FButton(
-                onPress: controller.toggle,
-                label: const Text('target'),
-              ),
+        TestScaffold.app(
+          child: FPopover(
+            controller: controller,
+            followerBuilder: (context, style, _) => const Text('follower'),
+            target: FButton(
+              onPress: controller.toggle,
+              label: const Text('target'),
             ),
           ),
         ),
@@ -38,20 +37,15 @@ void main() {
     });
 
     testWidgets('tap outside does not hide popover', (tester) async {
-      final controller = FPopoverController(vsync: const TestVSync());
-
       await tester.pumpWidget(
-        MaterialApp(
-          home: TestScaffold(
-            theme: FThemes.zinc.light,
-            child: FPopover(
-              controller: controller,
-              hideOnTapOutside: false,
-              followerBuilder: (context, style, _) => const Text('follower'),
-              target: FButton(
-                onPress: controller.toggle,
-                label: const Text('target'),
-              ),
+        TestScaffold.app(
+          child: FPopover(
+            controller: controller,
+            hideOnTapOutside: false,
+            followerBuilder: (context, style, _) => const Text('follower'),
+            target: FButton(
+              onPress: controller.toggle,
+              label: const Text('target'),
             ),
           ),
         ),
@@ -69,20 +63,15 @@ void main() {
     });
 
     testWidgets('tap button when popover is open closes it', (tester) async {
-      final controller = FPopoverController(vsync: const TestVSync());
-
       await tester.pumpWidget(
-        MaterialApp(
-          home: TestScaffold(
-            theme: FThemes.zinc.light,
-            child: FPopover(
-              controller: controller,
-              hideOnTapOutside: false,
-              followerBuilder: (context, style, _) => const Text('follower'),
-              target: FButton(
-                onPress: controller.toggle,
-                label: const Text('target'),
-              ),
+        TestScaffold.app(
+          child: FPopover(
+            controller: controller,
+            hideOnTapOutside: false,
+            followerBuilder: (context, style, _) => const Text('follower'),
+            target: FButton(
+              onPress: controller.toggle,
+              label: const Text('target'),
             ),
           ),
         ),
@@ -99,4 +88,34 @@ void main() {
       expect(find.text('follower'), findsNothing);
     });
   });
+
+  group('FPopover.tappable', () {
+    testWidgets('shown', (tester) async {
+      await tester.pumpWidget(
+        TestScaffold.app(
+          child: FPopover.tappable(
+            controller: controller,
+            followerBuilder: (context, style, _) => const Text('follower'),
+            target: Container(
+              color: Colors.black,
+              height: 10,
+              width: 10,
+            ),
+          ),
+        ),
+      );
+
+      await tester.tap(find.byType(Container).last);
+      await tester.pumpAndSettle();
+
+      expect(find.text('follower'), findsOneWidget);
+
+      await tester.tapAt(Offset.zero);
+      await tester.pumpAndSettle();
+
+      expect(find.text('follower'), findsNothing);
+    });
+  });
+
+  tearDown(() => controller.dispose());
 }