+++ forui/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Add `FBreakpoints`.
 
+* Add `FIcon.empty()`.
+
 * Add `FTappable`.
 
 * Add `FTile`.
@@ -16,11 +18,28 @@
 
 * Add `FSelectTileGroup`.
 
+* Add `FCalendarDayPickerStyle.tileSize`.
+
+* Add `FPopover.ignoreDirectionalPadding`.
+
+* Add `FPopover.tappable(...)`.
+
+* Add `FPopoverMenu`.
+
+* Add `FPortal.offset`.
+
+### Changes
+
+* **Breaking** Change `FPopover()` to not automatically wrap a target in a `GestureDetector` - use `FPopover.tappable(...)`
+  instead.
+
 ### Fixes
 
 * Change FButton's animation to only start on mouse down and up.
 
 * Fix `FLabel` not showing error message if label and description are null.
+
+* Fix `FSelectGroup` not properly disposing callbacks.
 
 
 ## 0.6.1